requirements.txt
Flask
mysql-connector-python
python-dotenv
Gunicorn
flask_cors